refactor: 컨트롤러 네이밍 변경 및 url 수정
