+++ src/main/java/org/prgms/kdtspringweek1/controller/thymeleafController/ThymeleafVoucherController.java
@@ -13,6 +13,7 @@ import java.util.List;
 import java.util.UUID;
 
 @Controller
+@RequestMapping("thymeleaf")
 @Profile({"thymeleaf"})
 public class ThymeleafVoucherController {
